+++ geotrek/feedback/models.py
@@ -16,9 +16,8 @@ from django.template.loader import render_to_string
 from django.utils import timezone
 from django.utils.formats import date_format
 from django.utils.translation import gettext_lazy as _
-from mapentity.models import MapEntityMixin
 
-from geotrek.common.mixins.models import AddPropertyMixin, NoDeleteMixin, PicturesMixin, TimeStampedModelMixin
+from geotrek.common.mixins.models import AddPropertyMixin, NoDeleteMixin, PicturesMixin, TimeStampedModelMixin, GeotrekMapEntityMixin
 from geotrek.common.utils import intersecting
 from geotrek.core.models import Path
 from geotrek.trekking.models import POI, Service, Trek
@@ -107,7 +106,7 @@ class PendingSuricateAPIRequest(models.Model):
         super().delete(*args, **kwargs)
 
 
-class Report(MapEntityMixin, PicturesMixin, TimeStampedModelMixin, NoDeleteMixin, AddPropertyMixin, ZoningPropertiesMixin):
+class Report(GeotrekMapEntityMixin, PicturesMixin, TimeStampedModelMixin, NoDeleteMixin, AddPropertyMixin, ZoningPropertiesMixin):
     """User reports, submitted via *Geotrek-rando* or parsed from Suricate API."""
 
     email = models.EmailField(verbose_name=_("Email"))