+++ geotrek/maintenance/models.py
@@ -11,19 +11,22 @@ from django.utils.translation import gettext_lazy as _
 
 from geotrek.altimetry.models import AltimetryMixin
 from geotrek.authent.models import StructureRelated, StructureOrNoneRelated
-from geotrek.common.mixins.models import TimeStampedModelMixin, NoDeleteMixin, AddPropertyMixin
+from geotrek.common.mixins.models import (TimeStampedModelMixin, NoDeleteMixin, AddPropertyMixin,
+                                          GeotrekMapEntityMixin, get_uuid_duplication)
 from geotrek.common.models import Organism
 from geotrek.common.utils import classproperty
 from geotrek.core.models import Topology, Path, Trail
 from geotrek.maintenance.managers import InterventionManager, ProjectManager
 from geotrek.zoning.mixins import ZoningPropertiesMixin
-from mapentity.models import MapEntityMixin
+
+from mapentity.models import DuplicateMixin
+
 
 if 'geotrek.signage' in settings.INSTALLED_APPS:
     from geotrek.signage.models import Blade
 
 
-class Intervention(ZoningPropertiesMixin, AddPropertyMixin, MapEntityMixin, AltimetryMixin,
+class Intervention(ZoningPropertiesMixin, AddPropertyMixin, GeotrekMapEntityMixin, AltimetryMixin,
                    TimeStampedModelMixin, StructureRelated, NoDeleteMixin):
 
     target_type = models.ForeignKey(ContentType, null=True, on_delete=models.CASCADE)
@@ -69,6 +72,10 @@ class Intervention(ZoningPropertiesMixin, AddPropertyMixin, MapEntityMixin, Alti
     objects = InterventionManager()
 
     geometry_types_allowed = ["LINESTRING", "POINT"]
+
+    elements_duplication = {
+        "attachments": {"uuid": get_uuid_duplication}
+    }
 
     class Meta:
         verbose_name = _("Intervention")
@@ -367,7 +374,7 @@ class InterventionJob(StructureOrNoneRelated):
         return self.job
 
 
-class ManDay(models.Model):
+class ManDay(DuplicateMixin, models.Model):
 
     nb_days = models.DecimalField(verbose_name=_("Mandays"), decimal_places=2, max_digits=6)
     intervention = models.ForeignKey(Intervention, on_delete=models.CASCADE)
@@ -385,7 +392,7 @@ class ManDay(models.Model):
         return str(self.nb_days)
 
 
-class Project(ZoningPropertiesMixin, AddPropertyMixin, MapEntityMixin, TimeStampedModelMixin,
+class Project(ZoningPropertiesMixin, AddPropertyMixin, GeotrekMapEntityMixin, TimeStampedModelMixin,
               StructureRelated, NoDeleteMixin):
 
     name = models.CharField(verbose_name=_("Name"), max_length=128)
@@ -411,6 +418,10 @@ class Project(ZoningPropertiesMixin, AddPropertyMixin, MapEntityMixin, TimeStamp
     eid = models.CharField(verbose_name=_("External id"), max_length=1024, blank=True, null=True)
 
     objects = ProjectManager()
+
+    elements_duplication = {
+        "attachments": {"uuid": get_uuid_duplication}
+    }
 
     class Meta:
         verbose_name = _("Project")
@@ -617,7 +628,7 @@ class Contractor(StructureOrNoneRelated):
         return self.contractor
 
 
-class Funding(models.Model):
+class Funding(DuplicateMixin, models.Model):
 
     amount = models.FloatField(verbose_name=_("Amount"))
     project = models.ForeignKey(Project, verbose_name=_("Project"), on_delete=models.CASCADE)