+++ geotrek/trekking/tests/test_views.py
@@ -25,8 +25,8 @@ from geotrek.authent.tests.base import AuthentFixturesTest
 from geotrek.authent.tests.factories import TrekkingManagerFactory, StructureFactory, UserProfileFactory
 from geotrek.common.templatetags import geotrek_tags
 from geotrek.common.tests import CommonTest, CommonLiveTest, TranslationResetMixin, GeotrekAPITestCase
-from geotrek.common.tests.factories import (AttachmentFactory, ThemeFactory, LabelFactory,
-                                            RecordSourceFactory, TargetPortalFactory)
+from geotrek.common.tests.factories import (AttachmentFactory, ThemeFactory, LabelFactory, RecordSourceFactory,
+                                            TargetPortalFactory)
 from geotrek.common.utils.testdata import get_dummy_uploaded_image
 from geotrek.core.tests.factories import PathFactory
 from geotrek.infrastructure.models import Infrastructure
@@ -342,7 +342,7 @@ class TrekViewsTest(GeotrekAPITestCase, CommonTest):
             'duration': 1.5,
             'id': self.obj.pk,
             'name': self.obj.name_display,
-            'thumbnail': 'None'
+            'thumbnail': 'None',
         }
 
     def get_bad_data(self):
@@ -1457,7 +1457,7 @@ class ServiceViewsTest(GeotrekAPITestCase, CommonTest):
     def get_expected_datatables_attrs(self):
         return {
             'id': self.obj.pk,
-            'name': self.obj.name_display
+            'name': self.obj.name_display,
         }
 
     def get_good_data(self):